backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
pydantic[email]
python-jose[cryptography]
bcrypt==4.0.1
passlib[bcrypt]
python-multipart
google-genai
langchain
langchain-community
langchain-core
langchain-ollama
chromadb
ollama
pypdf
docx2txt
python-dotenv
pydantic-settings
